# fileupload_batch_swagger/swagger.yaml
swagger: "2.0"
info:
  description: "FeedBackManagementSystem - Using Spring 5 Reactive programming and React JS "
  version: "1.0.0"
  title: "FMS-Outreach Events"
...
paths:
...
  /generateReports:
    get:
      tags:
      - "Admin"
      summary: "In Reports Screen, Admin can see the reports for events based on any filter criteria and email the filtered report to BU heads"
      description: ""
      operationId: "generateReports"
      produces:
      - "application/json"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/GenerateReports"
        400:
          description: "Invalid status value"
...
  /addFeedbackQuestions{questionType}:
    get:
      tags:
      - "Admin"
      summary: "Admin can add/edit feedback questions in real-time."
      description: ""
      operationId: "addFeedbackQuestions"
      produces:
      - "application/json"
      parameters:
      - name: "questionType"
        in: "path"
        description: "The name that needs to be fetched. Use user1 for testing. "
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/FeedbackQuestions"
        400:
          description: "Invalid status value"
    post:
      tags:
      - "Admin"
      summary: "Admin can add feedback questions in real-time"
      operationId: "addQuestion"
      produces:
      - "application/json"
      parameters:
      - name: "questionType"
        in: "path"
        description: "The Employee that needs to be fetched. Use Employee for testing."
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/AddQuestion"
        400:
          description: "Invalid status value"
    put:
      tags:
      - "Admin"
      summary: "Admin can edit feedback questions in real-time"
      description: "This can only be done by the Admin Role."
      operationId: "editQuestion"
      produces:
      - "application/json"
      parameters:
      - name: "questionType"
        in: "path"
        description: "The EmployeeId that needs to be removed/Deleted"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/EditQuestion"
        400:
          description: "Invalid status value"
...
definitions:
...
  GenerateReports:
    type: "object"
    properties:
      eventId:
        type: "string"
      baseLocation:
        type: "string"
      beneficiaryName:
        type: "string"
      councilName:
        type: "string"
      eventName:
        type: "string"
      eventDescription:
        type: "string"  
      eventDate:
        type: "string"
        format: "date-time"
      employeeId:
        type: "string"    
      employeeName:
        type: "string"    
      volunteerHours:
        type: "integer"
        format: "int32"
      travelHours:
        type: "integer"
        format: "int32"
      livesImpacted:
        type: "integer"
        format: "int32"
      businessUnit:
        type: "string"
      status:
        type: "string"
      IIEPCategory:
        type: "string"
...
  FeedbackQuestions:
    type: "object"
    properties:
      questions:
        type: "string"
      totalAnswers:
        type: "integer"
        format: "int32"
      feedBackType:
        type: "string"
  AddQuestion:
    type: "object"
    properties:
      participated:
        type: "string"
      notParticipated:
         type: "string"
      unregistered:
        type: "string"
      allowmulitpleAnswers:
         type: "string"
      freeTextAnswer:
         type: "string"
      customQuestion:
         type: "string"
  EditQuestion:
    type: "object"
    properties:
      participated:
        type: "string"
      notParticipated:
         type: "string"
      unregistered:
        type: "string"
      allowmulitpleAnswers:
         type: "string"
      freeTextAnswer:
         type: "string"
      customQuestion:
         type: "string"
...